+++ tools/jreq/jreq.java
@@ -17,58 +17,46 @@
 	along with JREq. If not, see <http://www.gnu.org/licenses/>.
 */
 
-class JREqDecl {
- protected static String affix = ""; // We are not interested in a component.
-}
+// COMPILE WITH OPTION '-d .' : 'javac -d . *.java'.
 
+package info.q37.jreq;
 
-// Begin of generic part.
-class JREq extends JREqDecl {
+class Wrapper {
 	native public static String wrapperInfo();
-	native public static String componentInfo( long launcher );
-	native private static void init( String location);
-	native private static long register( String arguments );
-	native protected static Object wrapper(
-		long launcher,
-		int index,
-		Object... objects);
 
-	static
+	native protected static String componentInfo(long launcher);
+
+	native protected static void init(String location);
+
+	native protected static long register(String arguments);
+
+	native protected static Object wrap(long launcher, int index, Object... objects);
+	static protected String getLocation()
 	{
-  String location = "";
-  String osName = System.getProperty("os.name").toLowerCase();
+		String osName = System.getProperty("os.name").toLowerCase();
 
- 	System.loadLibrary( "jreq" );
+		if (System.getenv("EPEIOS_SRC") == null) {
+			return "./";
+		} else if (osName.contains("windows"))
+			return "h:/bin/";
+		else if (osName.contains("mac")) {
+			return  "/Users/csimon/bin/";
+		} else {
+			return "/home/csimon/bin/";
+		}
 
-  if ( osName.contains( "windows" ) )
-   location = "h:/bin";
-  else if ( osName.contains( "mac" ) ) {
-   location = "/Users/bin";
-  } else {
-   location = "/home/csimon/bin";
-  }
-
-  if ( System.getenv( "EPEIOS_SRC" ) == null ) {
-   location = ".";
-  }
-
- 	init( location );
-//  register( "./" + JREqDecl.affix + "jre");
-  }
-}
-// End of generic part.
-
-
-class JREqTest {
-	private static void displayCompilationTime() throws Exception
-	{
-		System.out.println( "Bytecode build : " + new java.util.Date(new java.io.File(JREqTest.class.getClassLoader().getResource(JREqTest.class.getCanonicalName().replace('.', '/') + ".class").toURI()).lastModified()) );
 	}
-	
-	public static void main ( String[] args ) throws Exception
-	{
- 	System.out.println( JREq.wrapperInfo() );
- 	displayCompilationTime();
- }
+	static {
+		String location = getLocation();
+
+		System.loadLibrary("jreq");
+
+		init(location);
+	}
 }
 
+public class JREq extends Wrapper {
+	static public long register( String componentName ) {
+		return Wrapper.register( getLocation() + componentName + "jre" );
+	}
+}